requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.1
langgraph>=0.0.11
langchain-mcp-adapters>=0.0.2
langchain-core>=0.1.0
python-dotenv>=1.0.0
typing-extensions>=4.5.0
requests>=2.31.0 